refactor(mod-lock): レビュー指摘対応（integrity厳格化・collectModIds共有化・yaml依存修正ほか）
コードレビュー指摘のうち妥当なものを反映:
- loader: yaml を devDependencies→dependencies へ（cli.ts の実行時依存として正しい）。
- shared: IntegritySchema を sha256 base64 の固定長（43文字+`=`）に厳格化。`sha256-AAAA` 等の
  短い不正値を弾く。実 lock（default.lock.json）で検証済み。
- shared: collectModIds を純関数として切り出し、loader.buildWorldLock と backend.collectMods の
  重複走査を一本化。
- loader: worker の content-type ガードのコメントを明確化（JS/text-plain/空 のみ許可、
  html/json 等は拒否）。module キャッシュが TTL を持たない設計意図（version キー＋no-store）も明記。
- frontend: buildWorldLock で lock 化できなかった mod を検出したとき警告（外部公開時に
  ロード拒否になる旨）。ローカルは寛容なのでブロックはしない。
- loader: acquireMod のテスト追加（lockEntry はあるが対象 component が lock に無い外部 → lock-missing）。

却下: (1) 既存 _migrateLegacyDbRecords の let は本PR差分外かつ await 逐次副作用ループの
accumulator で関数型化はむしろ複雑化するため据え置き。(2) data-only の progress は
完了で相殺され欠陥ではないため据え置き。
